requirements.txt
asgiref==3.9.1
attrs==25.3.0
channels==4.1.0
channels-redis==4.2.0
coverage==7.6.0
Django==5.0.6
django-stubs==5.2.2
django-stubs-ext==5.2.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
msgpack==1.1.1
mypy==1.17.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==10.4.0
pluggy==1.6.0
psycopg==3.2.9
psycopg-binary==3.2.9
pytest==8.2.0
pytest-django==4.8.0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.7
referencing==0.36.2
rpds-py==0.27.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20250822
typing_extensions==4.14.1
uritemplate==4.2.0
whitenoise==6.6.0
